affichage des jobs

+++ GdcServerManager/src/dao/project/MySQLJobDAO.java
@@ -12,7 +12,6 @@ import settings.sql.DBTables;
 import settings.sql.tables.JobTable;
 
 
-
 import dao.MySQLUserDAO;
 import dao.UserDAO;
 
@@ -20,6 +19,38 @@ import model.User;
 import model.Job;
 
 public class MySQLJobDAO implements JobDAO {
+	public ArrayList<Job> retrieveAllJob() throws SQLException {
+		ArrayList<Job> jobs = new ArrayList<Job>();
+		ResultSet rset = null;
+		Statement stmt = null;
+		Connection connection = null;
+		try {
+			connection = SQLSettings.getPDS().getConnection();
+			UserDAO userdao=new MySQLUserDAO();	
+			stmt = connection.createStatement();
+			
+			rset = stmt.executeQuery("select * from "+SQLSettings.TABLES.getJob().TNAME+"");
+
+			// boucle sur les resultats de la requête
+			while (rset.next()) {
+				Job job = new Job();
+				job.setId(rset.getInt(SQLSettings.TABLES.getJob().getId()));
+				job.setUserId(userdao.retrieveUser(rset.getInt(SQLSettings.TABLES.getJob().getUserId())));
+				job.setJobId(rset.getString(SQLSettings.TABLES.getJob().getJobId()));
+				job.setOs(rset.getString(SQLSettings.TABLES.getJob().getOs()));
+				job.setDescription(rset.getString(SQLSettings.TABLES.getJob().getDescription()));
+				jobs.add(job);
+			}
+			return jobs;
+		} catch (Exception e) {
+			e.printStackTrace();
+			throw e;
+		} finally {
+			try { if(rset!=null) rset.close();  } catch (Exception e) {};
+			try { if(stmt!=null) stmt.close();  } catch (Exception e) {};
+			try { if(connection!=null) connection.close();  } catch (Exception e) {};
+		}
+	}
 
 	public boolean newJob(int user_id,String jobid,String submitDate,String os,String description) throws SQLException {
 
@@ -29,7 +60,6 @@ public class MySQLJobDAO implements JobDAO {
 		try {
 			connection = SQLSettings.getPDS().getConnection();
 			stmt = connection.createStatement();
-
 			rset = stmt.execute("insert into "+SQLSettings.TABLES.getJob().TNAME+" values (NULL,"
 					+ user_id + ",'"+jobid+"','"+submitDate+"','"+os+"','"+description+"')");
 
@@ -156,7 +186,7 @@ public class MySQLJobDAO implements JobDAO {
 
 	}
 	
-	public void removeJob(String jobid) throws SQLException {
+	public void removeJob(String jobid, int user_id) throws SQLException {
 		int rset = 0;
 		Statement stmt = null;
 		Connection connection = null;
@@ -166,8 +196,8 @@ public class MySQLJobDAO implements JobDAO {
 			
 			DBTables tab = SQLSettings.TABLES;
 			JobTable nt = tab.getJob();	
-
-			rset = stmt.executeUpdate("delete from "+nt.TNAME+" where "+nt.getJobId()+"=" +jobid);
+			
+			rset = stmt.executeUpdate("delete from "+nt.TNAME+" where "+nt.getJobId()+"=" +jobid+"and "+nt.getUserId()+"="+user_id);
 			
 			return;
 		} catch (SQLException e2) {
